to long

+++ Platipus.Wallet.Api/Application/Requests/Wallets/Openbox/Base/Response/PswBalanceResponse.cs
@@ -1,5 +1,5 @@
 namespace Platipus.Wallet.Api.Application.Requests.Wallets.Openbox.Base.Response;
 
-public record OpenboxBalanceResponse(decimal Balance);
+public record OpenboxBalanceResponse(long Balance);
 
 public record OpenboxTokenResponse(Guid Token);
